"Rearranged delUserOutreaches cloudfunction"
